+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - ReleaseDate
 
+### Changed
+
+- Name of header to Merging Branches.
+- Branches in table now in bold style.
+
 ## [v0.2.2] - 2023-05-06
 
 ### Fixed
@@ -59,9 +64,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Output is given in terminal application as a table.
 
 <!-- next-url -->
+
 [Unreleased]: https://github.com/strowk/probranchinator/compare/v0.2.2...HEAD
 [v0.2.2]: https://github.com/strowk/probranchinator/compare/v0.2.2...v0.2.2
-
 [v0.2.2]: https://github.com/strowk/probranchinator/compare/v0.2.1...v0.2.2
 [v0.2.1]: https://github.com/strowk/probranchinator/compare/v0.2.0...v0.2.1
 [v0.2.0]: https://github.com/strowk/probranchinator/compare/v0.1.0...v0.2.0